+++ src/core/judgingthread.h
@@ -41,7 +41,7 @@ class JudgingThread : public QThread {
 	void setRawMemoryLimit(int);
 	void setInterpreterAsWatcher(bool);
 	int getTimeUsed() const;
-	int getMemoryUsed() const;
+	qint64 getMemoryUsed() const;
 	int getScore() const;
 	int getFullScore() const;
 	int getJudgeTimes() const;
@@ -72,7 +72,7 @@ class JudgingThread : public QThread {
 	int memoryLimit{};
 	int rawMemoryLimit{};
 	int timeUsed;
-	int memoryUsed;
+	qint64 memoryUsed;
 	int score{};
 	int judgedTimes;
 	ResultState result;
