+++ src/detaildialog.cpp
@@ -133,7 +133,7 @@ void DetailDialog::refreshViewer(Contest *_contest, Contestant *_contestant) {
 		QList<QList<ResultState>> result = contestant->getResult(i);
 		QList<QStringList> message = contestant->getMessage(i);
 		QList<QList<int>> timeUsed = contestant->getTimeUsed(i);
-		QList<QList<int>> memoryUsed = contestant->getMemoryUsed(i);
+		QList<QList<qint64>> memoryUsed = contestant->getMemoryUsed(i);
 		QList<QList<int>> score = contestant->getScore(i);
 
 		for (int j = 0; j < inputFiles.size(); j++) {
