+++ src/judgingdialog.cpp
@@ -83,7 +83,7 @@ void JudgingDialog::judgeAll() {
 }
 
 void JudgingDialog::singleCaseFinished(QString contestantName, int progress, int x, int y, int result,
-                                       int scoreGot, int timeUsed, int memoryUsed) {
+                                       int scoreGot, int timeUsed, qint64 memoryUsed) {
 	bool isOnMaxValue =
 	    ui->logViewer->verticalScrollBar()->value() == ui->logViewer->verticalScrollBar()->maximum();
 	QTextBlockFormat blockFormat;
